pyproject.toml
[project]
name = "MoMaRestServer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "fastapi>=0.124.0",
    "momafksolverproject",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]

[tool.uv.sources]
momafksolverproject = { path = "../MoMa_FK_Solver", editable = true }
